requirements.txt
matplotlib==3.1.3
torch==1.4.0
torchvision==0.5.0
numpy==1.18.1
cvxpy==1.0.28
pycddlib==2.1.1
onnxruntime==1.2.0
onnx_simplifier==0.2.7
pandas==1.0.2
Keras==2.3.1
tqdm==4.43.0
scipy==1.3.1
cleverhans==3.0.1
setproctitle==1.1.10
gpustat==0.5.0
onnx==1.5.0
numba==0.48.0
cdd==0.1.12
gurobipy==5.0.2
Pillow==7.1.2
posix_ipc==1.0.4
setGPU==0.0.7
tensorflow-gpu==1.15.2
waitGPU==0.0.3